adaptix/services/inventory/requirements.txt
django>=4.2,<5.0
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
gunicorn
python-jose[cryptography]
requests
django-cors-headers
dj-database-url
PyJWT>=2.8.0
celery
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-django
opentelemetry-instrumentation-psycopg2
opentelemetry-instrumentation-requests
opentelemetry-exporter-otlp
pytest-django
whitenoise
django-filter>=23.2
